+++ apps/inspector/src/components/siteProfile/index.ts
@@ -1,3 +1,2 @@
 export * from "./SiteProfile";
 export * from "./use-site-profile";
-export * from "./verify-website";
