+++ packages/extension-common/src/site-profile/index.ts
@@ -1,3 +1,4 @@
 export * from "./events";
 export * from "./messaging";
 export * from "./types";
+export * from "./verify-website";
