+++ packages/extension-common/src/utils/warning-params.ts
